rename delete study site modal
